GameAdivinha

+++ src/OutroLoop.java
@@ -5,7 +5,7 @@ public class OutroLoop {
         Scanner leitura = new Scanner(System.in);
         int mediaAvaliacao = 0;
         double nota = 0;
-        int totalDeNotas = 0;
+        double totalDeNotas = 0;
 
         while (nota != -1){
             System.out.println("Digite a sua avaliação ou -1 para encerrar: ");
